+++ totalRP3_Storyline/trunk/changelog.md
@@ -1,5 +1,6 @@
-# Change-log version 1.2.1
+# Change-log version 1.2.2
 
-## Bug fixes
+## UI improvements
 
-- Fixed the issue with some models that were not placed correctly. Sorry about that.
+- We are now using a **new system to hide the default quest frame and dialog frame**. The frames should no longer re-appear when opening other frames, and the other frames should be correctly aligned to the left of the screen without any gap. This new solution require a reload of the user interface.
+- With the new system to hide the quest frame, we also implement an **option to use the default interface layout engine**. When opening default frames, they appear on the left of the screen, one next to the other, with priorities. Storyline can now use that layout engine to be placed just like the character pannel or spellbook.